requirements.txt
dj-database-url==0.5.0
Django==2.1
django-import-export==1.1.0
django-tables2==2.0.3
gunicorn==19.9.0
tables==3.4.3
tablib==0.12.1
tblib==1.3.2
django-crispy-forms==1.7.2
psycopg2-binary==2.7.6
whitenoise==4.1.2